+++ tests/Integration/InitializationTest.php
@@ -13,10 +13,7 @@ use PHPUnit\Framework\TestCase;
 class InitializationTest extends TestCase {
 
 	public function testVersionConstantIsDefined() {
-		$this->assertIsString( Maps_VERSION );
-		$this->assertIsString( SM_VERSION );
-		$this->assertSame( Maps_VERSION, SM_VERSION );
-		$this->assertNotEmpty( Maps_VERSION );
+		$this->assertIsInt( NS_GEO_JSON );
 	}
 
 }